Update schema.sql
The existing BjyAuthorize\Provider\Role\ZendD configuration outlined in the readme.md expects the role_id in the user_role_linker table to be the varchar role_id user in the user_role table not the id of the user_role. Because of this using the schema.sql provided causes errors to be thrown. This fix changes the data type of role_id column to varchar as the provider is expecting.